separation crawler
