+++ .gitignore
@@ -1,2 +1,3 @@
 /logs/**
 /node_modules
+/dist
