+++ package.json
@@ -4,6 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
+    "build": "tsc",
     "start": "node index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -11,7 +12,12 @@
   "license": "ISC",
   "dependencies": {
     "dayjs": "^1.11.7",
+    "hangul-js": "^0.2.6",
     "puppeteer": "^20.4.0",
     "readline": "^1.3.0"
+  },
+  "devDependencies": {
+    "@types/puppeteer": "^7.0.4",
+    "typescript": "^5.5.4"
   }
 }
