+++ src/types/korail-station-name.interface.ts
@@ -1,0 +1,3 @@
+export interface KorailStaionName extends Element {
+  innerText: string;
+}
