+++ src/types/page-branding.type.ts
@@ -1,0 +1,5 @@
+import { Page } from "puppeteer"
+
+type Brand<Key extends string, value> = value & { __brand: Key }
+
+export type LoginedPage = Brand<'logined', Page>;
