+++ src/types/station-name-branding.type.ts
@@ -1,0 +1,3 @@
+type Brand<Key extends string, value> = value & { __brand: Key };
+
+export type StationName = Brand<'stationName', string>;
